requirements.txt
numpy
pillow
pywavelets
opencv-python
scipy
